Fix bug with long usernames
Long usernames were causing the width to stretch in mobile view on DCR
due to the width not being set on a container. This moves from using a
`Column` for comments, to a div with the same flex styling but with a
`max-width` of `100%`.

I've also added a story to explicitly check for threaded replies with
long usernames, though in this case that wasn't causing an issue as it
was a combination of factors that only occurred when implementing in
DCR.