Update logic for handling contributors
After testing it was found that the badge array will contain an element
for Contributors, similarly for Staff. Given that I have a badge array,
I'm assuming I could get multiple badges and so therefore still want to
handle the case where a user is marked as both Staff and Contributor,
but we only show the Staff badge.